Implement `Bid` encoding/compression functions
We need to store Bids in a tree. But for ZKProofs
purposes and due to the usage of Poseidon Hashing
algorithm, we can only hash bls12_381::Scalar.

To solve this problem, dmitry sent a document on which
is explained a cannonical way to encode this.
See:  https://hackmd.io/@7dpNYqjKQGeYC7wMlPxHtQ/BkfS78Y9L

- Implement error handling for encoding failures.
- Implement encoding functions.

This is part of the work needed on #5