Add encoding gadget functions
Currently, we need to print into the CS the way on which
we compute the encoded leaf value for a Bid.

To do so, we need to perform the poseidon sponge function inside
of the circuit and vinculate the inputs used later on.

- Create functions to perform the gadget sponge hash function after
the complete encoding of the bid.
- Added collapsing function that prints the sponge hash into the
constraint system.
- Implemented tests to check whether the gadget and the `Scalar`
method provide the same results.