Refactor Bid encoding as TryInto impl
As defined on https://github.com/dusk-network/Poseidon252/pull/28
we will need to express the encoding procedures for our complex
data structures as an `Into` or `TryInto` impl.

Therefore, this PR is addressed to refactor the Bid encoding procedure
we had implemented on this repo to work accordingly with the new
PoseidonTree requirements.

Closes #12