Implement StorageBid structure
This is required by #14. Since we need to store Bids, but not all of
them fields are required, we implemented `StorageBid` which is the
structure that will be stored inside of the Bid Contract Tree.

We will need to provide a `From` impl for the Bid and also refactor
the encoding procedure to work with the `StorageBid` instead of the
`Bid` itself.