Replace pub inputs by constrained witnesses
In order to make it easier to elaborate the gadgets
and work with the Public Inputs, we created them as witnesses
constrained to be equal to a publicly known constant.

This solved the errors that we had on the tests since we weren't
passing any Public Inputs vector to the verification fn.

Closes #27