Fix `complex_rangeproof var_len circuits bug
We realized that the circuits that the `blindbidproof gadget`
was generating had different circuit sizes.

After researching a bit, we find out that the accumulator
computation (where we check that the value is indeed equal
to the original one) was iterating over `bit_count` instead
of a fixed sice. And therefore, this caused the circuits to
have a different ammount of constraints.

Fixed the bug by iterating over the 256-bit representation of the
number and seting the `bit = 0` if `idx >= num_bits`.