Refactor circuits & impl to adapt to ElGamal
We've edited the circuits and the implementation in
order to adapt the repo to ElGamal encryption schema.

Both the value and the blinder fields are now provided to
the gadgets but ASAP once the ElGmal schema is implemented
they will be decrypted and used instead of being provided.

Apart from that, the versions of the dependencies have been
updated and restrictions to the Bid generation have been added
in order to not surpass the maximum bid value allowed.