Replace ctant witnesses per normal ones
As mentioned in #38 That's actually a problem since we
define `value` and `bid_value` as constant witnesses
(constraint to a constant).
This means that we're harcoding these values in the
`ProverKey` & `VerifierKey` and therefore,
the circuit is not reusable.

Solved it by assigning a normal witness inclusion for these
values.