Refactor the BidGenerationError enum with lifetime
Actually I tried to avoid the need of copying or cloning the
`JubJubScalar` when we sent it to the `Error`.

The problem is that since the `Error` returned is not behind
a reference, the lifetime required would be `'static` and
that's not possible to achieve. The quickest solution is to copy
the `found` parameter and get a `'static` ref to the `max_val`
& `min_val`.