Define constants at module level instead of compute
We need to constantly operate with two pre-computed
numbers which do always have the same value.

Since they were coputed through two function calls the
compiler might not be able to optimize them and set them
directly as constants.
Therefore, we did it and now they are declared as constants
in the score_gen module.