Fix plonk_gadgets dependency in Cargo.toml
Me, @CPerezz prematurely merged #49 with a git dependency inside
that wasn't pointing to a tag/release. This is a bad practice and indeed
it can cause problems since now the branch refered in Cargo.toml
doesn't exist.

Therefore, now that we have the release, this PR updates the
`plonk_gadgets` dependency of ´dusk-blindbid` to the newest
tag `v0.2.0`.