Refactor BlindBidproof gadget & outsource tests
The blindbid gadget wasn't having the appropiate fn
signatures and wasn't using the single Bid structure
approach introduced on the previous commits.
Also, the tests weren't done as integration tests.

Therefore, the signature of the gadget fn has been refactored
to accept the consensus parameters and the score in order to
make the proof.
Also, we've moved the tests from the proof.rs file to the integration
tests folder.

This closes #56 & #51