Refactor bid_preimage gadget linking Variables
We were not linking the variables previously used by the
blindbid circuit with the variables used inside of the
bid_preimage_gadget fn. Indeed we were re-declaring them
which is not correct.

Now the function requires the fields directly as Variables
and we can link them to the ones used previously.

Also, the fn now returns the resulting hash so that it can
be checked against a PI in the circuit definition.