Remap decryption errors to return dummy values
If the decryption fails, we just set the result to an
impossible-to-obtain value.

On that way, verifiers do not get stuck on the process (they don't
care) about the real values here (just about filling the
composer). And provers won't get any info about if this
secret can or not decrypt the cipher.