Fix Variable-length circuit issues on round checks
We were experiencing proving problems in
https://github.com/dusk-network/rusk/pull/80 due to the fact that
we were doing rangeproofs with variant max ranges.

This was causing the proofs to fail when they shouldnt since the
ProverKey & VerifierKey were basically incorrect for the circuits
that the proofs were addressed for.

This has been solved and now the constraints are done against
constant max bounds by subtracting the witnesses.