Make BidTree `Canon` compatible
We need to be able to use `BidTree` structure inside of Contracts
together with `Canon`. And for this to be possible, we needed to
implement `Canon` and other traits required by `Poseidon252` &
`microkelvin` so that the functionalities of the tree  are the
expected ones.

- Updated the versions used of `canonical`, `poseidon252` &
`microkelvin`.
- Derived `Canon` for `BidTree`.
- Updated the API of the `BidTree`.
- Implement `PoseidonLeaf` & `Borrow<u64>` for `Bid`.
- Added `pos` field required by `Poseidon252` to the
`Bid` which exposes the position of the Bid in the tree.